requirements.txt
Django==1.11.1
django-environ==0.4.3
django-phonenumber-field==1.3.0
django-redis==4.8.0
django-rest-swagger==2.1.2
djangorestframework==3.6.3
djangorestframework-gis==0.11.2
drf-nested-routers==0.90.0
factory-boy==2.8.1
psycopg2==2.7.1
whitenoise==3.3.0